+++ include/mesh.hpp
@@ -29,7 +29,7 @@ public:
     std::vector<Vector3f> v;
     std::vector<TriangleIndex> t;
     std::vector<Vector3f> n;
-    bool intersect(const Ray &r, Hit &h, float t_min) override;
+    bool intersect(const Ray &r, Hit &h, double t_min) override;
 
 private:
     // Normal can be used for light estimation
